+++ src/components/multi-step-wizard/step/__spec__.js
@@ -164,6 +164,16 @@ describe('Step', () => {
         expect(instance.stepDisabled).toBeTruthy();
       });
     });
+
+    describe('when step is enabled', () => {
+      it('returns false', () => {
+        instance = TestUtils.renderIntoDocument(
+          <Step enabled={ true }>Demo Step</Step>
+        );
+        instance.context = stepContext;
+        expect(instance.stepDisabled).toBeFalsy();
+      });
+    });
   });
 
   describe('stepProcessed', () => {