+++ CHANGELOG.md
@@ -18,6 +18,9 @@
 ## Component Enhancements
 
 * `Decimal` component can prevent decimal value from exceeding the precision setting
+* `Step` wizard sub-component now accepts a prop of `enabled`.
+* `Table` components now accept an `onPageSizeChange` callback function as a prop.
+* `Pager` component now emits which element has changed.
 
 # 0.26.0